Refactor state management into slice modules
