+++ src/core/state.js
@@ -1,400 +1,216 @@
 import { DEFAULT_DAY_HOURS } from './constants.js';
-import { createId, structuredClone } from './helpers.js';
-import { attachRegistryMetricIds, buildMetricIndex } from '../game/schema/metrics.js';
-import nicheDefinitions from '../game/assets/nicheData.js';
+import { structuredClone } from './helpers.js';
 import {
   createEmptyCharacterState,
   createEmptySkillState,
   normalizeCharacterState,
   normalizeSkillState
 } from '../game/skills/data.js';
+import { ensureNicheStateShape } from './state/niches.js';
+import { normalizeAssetState } from './state/assets.js';
+import {
+  configureRegistry,
+  getRegistrySnapshot,
+  getHustleDefinition,
+  getAssetDefinition,
+  getUpgradeDefinition,
+  getMetricDefinition
+} from './state/registry.js';
 
-let registry = { hustles: [], assets: [], upgrades: [] };
-let hustleMap = new Map();
-let assetMap = new Map();
-let upgradeMap = new Map();
-let metricIndex = new Map();
-
-function getNicheIdSet() {
-  return new Set(nicheDefinitions.map(entry => entry.id));
-}
-
-function isValidNicheId(id) {
-  if (typeof id !== 'string' || !id) return false;
-  return getNicheIdSet().has(id);
-}
-
-function rollInitialNicheScore() {
-  const min = 25;
-  const max = 95;
-  const spread = max - min;
-  return Math.max(0, Math.min(100, Math.round(min + Math.random() * spread)));
-}
-
-function ensureNicheStateShape(target) {
-  target.niches = target.niches || {};
-  const nicheState = target.niches;
-  nicheState.popularity = nicheState.popularity || {};
-
-  const validIds = getNicheIdSet();
-  for (const id of Object.keys(nicheState.popularity)) {
-    if (!validIds.has(id)) {
-      delete nicheState.popularity[id];
-    }
+class StateManager {
+  constructor() {
+    this.state = null;
   }
 
-  for (const definition of nicheDefinitions) {
-    const entry = nicheState.popularity[definition.id] || {};
-    const score = Number(entry.score);
-    const previousScore = Number(entry.previousScore);
-    nicheState.popularity[definition.id] = {
-      score: Number.isFinite(score)
-        ? Math.max(0, Math.min(100, Math.round(score)))
-        : rollInitialNicheScore(),
-      previousScore: Number.isFinite(previousScore)
-        ? Math.max(0, Math.min(100, Math.round(previousScore)))
-        : null
+  createEmptyDailyMetrics() {
+    return {
+      time: {},
+      payouts: {},
+      costs: {}
     };
   }
 
-  const storedDay = Number(nicheState.lastRollDay);
-  nicheState.lastRollDay = Number.isFinite(storedDay) ? storedDay : target.day || 1;
-}
-
-export let state = null;
-
-export function createEmptyDailyMetrics() {
-  return {
-    time: {},
-    payouts: {},
-    costs: {}
-  };
-}
-
-export function ensureDailyMetrics(target = state) {
-  if (!target) return null;
-  target.metrics = target.metrics || {};
-  if (!target.metrics.daily) {
-    target.metrics.daily = createEmptyDailyMetrics();
-  }
-  return target.metrics.daily;
-}
-
-export function configureRegistry({ hustles = [], assets = [], upgrades = [] }) {
-  const prepared = attachRegistryMetricIds({ hustles, assets, upgrades });
-  registry = prepared;
-  hustleMap = new Map(prepared.hustles.map(item => [item.id, item]));
-  assetMap = new Map(prepared.assets.map(item => [item.id, item]));
-  upgradeMap = new Map(prepared.upgrades.map(item => [item.id, item]));
-  metricIndex = buildMetricIndex(prepared);
-}
-
-export function getHustleDefinition(id) {
-  return hustleMap.get(id);
-}
-
-export function getAssetDefinition(id) {
-  return assetMap.get(id);
-}
-
-export function getUpgradeDefinition(id) {
-  return upgradeMap.get(id);
-}
-
-export function getMetricDefinition(metricId) {
-  return metricIndex.get(metricId);
-}
-
-export function normalizeAssetInstance(definition, instance = {}) {
-  const normalized = { ...instance };
-  if (!normalized.id) {
-    normalized.id = createId();
-  }
-
-  const setupDays = Math.max(0, Number(definition?.setup?.days) || 0);
-  const status = normalized.status === 'active' || setupDays === 0 ? 'active' : 'setup';
-  normalized.status = status;
-
-  const remaining = Number(normalized.daysRemaining);
-  if (status === 'setup') {
-    normalized.daysRemaining = Number.isFinite(remaining) ? Math.max(0, remaining) : setupDays;
-  } else {
-    normalized.daysRemaining = 0;
-  }
-
-  const completed = Number(normalized.daysCompleted);
-  if (Number.isFinite(completed)) {
-    normalized.daysCompleted = Math.max(0, completed);
-  } else {
-    normalized.daysCompleted = status === 'active' ? setupDays : 0;
-  }
-
-  normalized.setupFundedToday = Boolean(normalized.setupFundedToday);
-  normalized.maintenanceFundedToday = Boolean(normalized.maintenanceFundedToday);
-
-  const usageEntries = Object.entries(normalized.dailyUsage || {});
-  const normalizedUsage = {};
-  for (const [key, value] of usageEntries) {
-    const numericValue = Number(value);
-    if (Number.isFinite(numericValue) && numericValue > 0) {
-      normalizedUsage[key] = Math.max(0, Math.floor(numericValue));
+  ensureDailyMetrics(target = this.state) {
+    if (!target) return null;
+    target.metrics = target.metrics || {};
+    if (!target.metrics.daily) {
+      target.metrics.daily = this.createEmptyDailyMetrics();
     }
+    return target.metrics.daily;
   }
 
-  normalized.dailyUsage = normalizedUsage;
-  delete normalized.cooldowns;
+  ensureStateShape(target = this.state) {
+    if (!target) return;
 
-  const lastIncome = Number(normalized.lastIncome);
-  normalized.lastIncome = Number.isFinite(lastIncome) ? lastIncome : 0;
-  const breakdown = normalized.lastIncomeBreakdown;
-  if (breakdown && typeof breakdown === 'object') {
-    const entries = Array.isArray(breakdown.entries)
-      ? breakdown.entries
-          .map(entry => {
-            if (!entry) return null;
-            const label = String(entry.label || '').trim();
-            const amount = Number(entry.amount);
-            if (!label || !Number.isFinite(amount)) return null;
-            return {
-              id: entry.id || null,
-              label,
-              amount: Math.round(amount),
-              type: entry.type || 'modifier',
-              percent: Number.isFinite(Number(entry.percent)) ? Number(entry.percent) : null
-            };
-          })
-          .filter(Boolean)
-      : [];
-    const total = Number(breakdown.total);
-    normalized.lastIncomeBreakdown = {
-      total: Number.isFinite(total) ? Math.max(0, Math.round(total)) : normalized.lastIncome,
-      entries
+    const registry = getRegistrySnapshot();
+
+    target.hustles = target.hustles || {};
+    for (const def of registry.hustles) {
+      const defaults = structuredClone(def.defaultState || {});
+      const existing = target.hustles[def.id];
+      target.hustles[def.id] = existing ? { ...defaults, ...existing } : defaults;
+    }
+
+    target.assets = target.assets || {};
+    for (const def of registry.assets) {
+      const existing = target.assets[def.id];
+      const normalized = normalizeAssetState(def, existing || {}, { state: target });
+      target.assets[def.id] = normalized;
+    }
+
+    target.upgrades = target.upgrades || {};
+    for (const def of registry.upgrades) {
+      const defaults = structuredClone(def.defaultState || {});
+      const existing = target.upgrades[def.id];
+      target.upgrades[def.id] = existing ? { ...defaults, ...existing } : defaults;
+      if (def.id === 'assistant') {
+        const assistantState = target.upgrades[def.id];
+        const storedCount = Number(assistantState.count);
+        if (!Number.isFinite(storedCount)) {
+          assistantState.count = assistantState.purchased ? 1 : 0;
+        } else {
+          assistantState.count = Math.max(0, storedCount);
+        }
+        if (assistantState.purchased && assistantState.count === 0) {
+          assistantState.count = 1;
+        }
+        delete assistantState.purchased;
+      }
+    }
+
+    target.totals = target.totals || {};
+    const earned = Number(target.totals.earned);
+    const spent = Number(target.totals.spent);
+    target.totals.earned = Number.isFinite(earned) && earned > 0 ? earned : 0;
+    target.totals.spent = Number.isFinite(spent) && spent > 0 ? spent : 0;
+
+    target.progress = target.progress || {};
+    target.progress.knowledge = target.progress.knowledge || {};
+
+    target.skills = normalizeSkillState(target.skills);
+    target.character = normalizeCharacterState(target.character);
+
+    this.ensureDailyMetrics(target);
+    ensureNicheStateShape(target, { fallbackDay: target.day || 1 });
+  }
+
+  buildBaseState() {
+    return {
+      money: 45,
+      timeLeft: DEFAULT_DAY_HOURS,
+      baseTime: DEFAULT_DAY_HOURS,
+      bonusTime: 0,
+      dailyBonusTime: 0,
+      day: 1,
+      hustles: {},
+      assets: {},
+      upgrades: {},
+      skills: createEmptySkillState(),
+      character: createEmptyCharacterState(),
+      totals: {
+        earned: 0,
+        spent: 0
+      },
+      progress: {
+        knowledge: {}
+      },
+      niches: {
+        popularity: {},
+        lastRollDay: 0
+      },
+      metrics: {
+        daily: this.createEmptyDailyMetrics()
+      },
+      log: [],
+      lastSaved: Date.now()
     };
-  } else {
-    normalized.lastIncomeBreakdown = null;
   }
-  const pendingIncome = Number(normalized.pendingIncome);
-  normalized.pendingIncome = Number.isFinite(pendingIncome) ? Math.max(0, pendingIncome) : 0;
-  const totalIncome = Number(normalized.totalIncome);
-  normalized.totalIncome = Number.isFinite(totalIncome) ? totalIncome : 0;
 
-  const createdOnDay = Number(normalized.createdOnDay);
-  normalized.createdOnDay = Number.isFinite(createdOnDay) ? Math.max(1, createdOnDay) : (state?.day ?? 1);
+  buildDefaultState() {
+    const base = this.buildBaseState();
+    this.ensureStateShape(base);
+    return base;
+  }
 
-  const quality = normalized.quality || {};
-  const level = Number(quality.level);
-  const normalizedLevel = Number.isFinite(level) ? Math.max(0, Math.floor(level)) : 0;
-  const progressEntries = Object.entries(quality.progress || {});
-  const normalizedProgress = {};
-  for (const [key, value] of progressEntries) {
-    const numericValue = Number(value);
-    if (Number.isFinite(numericValue) && numericValue > 0) {
-      normalizedProgress[key] = numericValue;
+  initializeState(initialState = null) {
+    this.state = initialState ? structuredClone(initialState) : this.buildDefaultState();
+    this.ensureStateShape(this.state);
+    return this.state;
+  }
+
+  replaceState(nextState) {
+    this.state = structuredClone(nextState);
+    this.ensureStateShape(this.state);
+    return this.state;
+  }
+
+  getState() {
+    return this.state;
+  }
+
+  getHustleState(id, target = this.state) {
+    target.hustles = target.hustles || {};
+    if (!target.hustles[id]) {
+      const def = getHustleDefinition(id);
+      target.hustles[id] = structuredClone(def?.defaultState || {});
     }
-  }
-  normalized.quality = {
-    level: normalizedLevel,
-    progress: normalizedProgress
-  };
-
-  const nicheId = typeof normalized.nicheId === 'string' ? normalized.nicheId : null;
-  normalized.nicheId = nicheId && isValidNicheId(nicheId) ? nicheId : null;
-
-  return normalized;
-}
-
-export function createAssetInstance(definition, overrides = {}) {
-  const setupDays = Math.max(0, Number(definition?.setup?.days) || 0);
-  const baseInstance = {
-    status: setupDays > 0 ? 'setup' : 'active',
-    daysRemaining: setupDays,
-    daysCompleted: setupDays > 0 ? 0 : setupDays,
-    setupFundedToday: false,
-    maintenanceFundedToday: false,
-    dailyUsage: {},
-    lastIncome: 0,
-    lastIncomeBreakdown: null,
-    pendingIncome: 0,
-    totalIncome: 0,
-    createdOnDay: state?.day ?? 1,
-    quality: {
-      level: 0,
-      progress: {}
-    },
-    nicheId: null
-  };
-  const merged = { ...baseInstance, ...structuredClone(overrides) };
-  if (merged.status === 'active') {
-    merged.daysRemaining = 0;
-    if (!Number.isFinite(Number(merged.daysCompleted))) {
-      merged.daysCompleted = setupDays;
-    }
-  }
-  return normalizeAssetInstance(definition, merged);
-}
-
-export function normalizeAssetState(definition, assetState = {}) {
-  const defaults = structuredClone(definition.defaultState || {});
-  const merged = { ...defaults, ...assetState };
-  if (!Array.isArray(merged.instances)) {
-    merged.instances = [];
+    return target.hustles[id];
   }
 
-  merged.instances = merged.instances.map(instance => normalizeAssetInstance(definition, instance));
-
-  if (merged.active && merged.instances.length === 0) {
-    merged.instances.push(createAssetInstance(definition, { status: 'active' }));
-  }
-
-  delete merged.active;
-  delete merged.buffer;
-  delete merged.fundedToday;
-
-  return merged;
-}
-
-export function ensureStateShape(target = state) {
-  if (!target) return;
-
-  target.hustles = target.hustles || {};
-  for (const def of registry.hustles) {
-    const defaults = structuredClone(def.defaultState || {});
-    const existing = target.hustles[def.id];
-    target.hustles[def.id] = existing ? { ...defaults, ...existing } : defaults;
-  }
-
-  target.assets = target.assets || {};
-  for (const def of registry.assets) {
-    const existing = target.assets[def.id];
-    const normalized = normalizeAssetState(def, existing || {});
-    target.assets[def.id] = normalized;
-  }
-
-  target.upgrades = target.upgrades || {};
-  for (const def of registry.upgrades) {
-    const defaults = structuredClone(def.defaultState || {});
-    const existing = target.upgrades[def.id];
-    target.upgrades[def.id] = existing ? { ...defaults, ...existing } : defaults;
-    if (def.id === 'assistant') {
-      const assistantState = target.upgrades[def.id];
-      const storedCount = Number(assistantState.count);
-      if (!Number.isFinite(storedCount)) {
-        assistantState.count = assistantState.purchased ? 1 : 0;
-      } else {
-        assistantState.count = Math.max(0, storedCount);
+  getAssetState(id, target = this.state) {
+    target.assets = target.assets || {};
+    const def = getAssetDefinition(id);
+    if (!def) {
+      if (!target.assets[id]) {
+        target.assets[id] = {};
       }
-      if (assistantState.purchased && assistantState.count === 0) {
-        assistantState.count = 1;
-      }
-      delete assistantState.purchased;
+      return target.assets[id];
     }
-  }
-
-  target.totals = target.totals || {};
-  const earned = Number(target.totals.earned);
-  const spent = Number(target.totals.spent);
-  target.totals.earned = Number.isFinite(earned) && earned > 0 ? earned : 0;
-  target.totals.spent = Number.isFinite(spent) && spent > 0 ? spent : 0;
-
-  target.progress = target.progress || {};
-  target.progress.knowledge = target.progress.knowledge || {};
-
-  target.skills = normalizeSkillState(target.skills);
-  target.character = normalizeCharacterState(target.character);
-
-  ensureDailyMetrics(target);
-  ensureNicheStateShape(target);
-}
-
-export function buildBaseState() {
-  return {
-    money: 45,
-    timeLeft: DEFAULT_DAY_HOURS,
-    baseTime: DEFAULT_DAY_HOURS,
-    bonusTime: 0,
-    dailyBonusTime: 0,
-    day: 1,
-    hustles: {},
-    assets: {},
-    upgrades: {},
-    skills: createEmptySkillState(),
-    character: createEmptyCharacterState(),
-    totals: {
-      earned: 0,
-      spent: 0
-    },
-    progress: {
-      knowledge: {}
-    },
-    niches: {
-      popularity: {},
-      lastRollDay: 0
-    },
-    metrics: {
-      daily: createEmptyDailyMetrics()
-    },
-    log: [],
-    lastSaved: Date.now()
-  };
-}
-
-export function buildDefaultState() {
-  const base = buildBaseState();
-  ensureStateShape(base);
-  return base;
-}
-
-export function initializeState(initialState = null) {
-  state = initialState ? structuredClone(initialState) : buildDefaultState();
-  ensureStateShape(state);
-  return state;
-}
-
-export function replaceState(nextState) {
-  state = structuredClone(nextState);
-  ensureStateShape(state);
-  return state;
-}
-
-export function getState() {
-  return state;
-}
-
-export function getHustleState(id, target = state) {
-  target.hustles = target.hustles || {};
-  if (!target.hustles[id]) {
-    const def = getHustleDefinition(id);
-    target.hustles[id] = structuredClone(def?.defaultState || {});
-  }
-  return target.hustles[id];
-}
-
-export function getAssetState(id, target = state) {
-  target.assets = target.assets || {};
-  const def = getAssetDefinition(id);
-  if (!def) {
-    if (!target.assets[id]) {
-      target.assets[id] = {};
-    }
+    target.assets[id] = normalizeAssetState(def, target.assets[id] || {}, { state: target });
     return target.assets[id];
   }
-  target.assets[id] = normalizeAssetState(def, target.assets[id] || {});
-  return target.assets[id];
-}
 
-export function countActiveAssetInstances(assetId, target = state) {
-  if (!assetId) return 0;
-  const assetState = getAssetState(assetId, target);
-  const instances = Array.isArray(assetState?.instances) ? assetState.instances : [];
-  return instances.filter(instance => instance?.status === 'active').length;
-}
-
-export function getUpgradeState(id, target = state) {
-  target.upgrades = target.upgrades || {};
-  if (!target.upgrades[id]) {
-    const def = getUpgradeDefinition(id);
-    target.upgrades[id] = structuredClone(def?.defaultState || {});
+  countActiveAssetInstances(assetId, target = this.state) {
+    if (!assetId) return 0;
+    const assetState = this.getAssetState(assetId, target);
+    const instances = Array.isArray(assetState?.instances) ? assetState.instances : [];
+    return instances.filter(instance => instance?.status === 'active').length;
   }
-  return target.upgrades[id];
+
+  getUpgradeState(id, target = this.state) {
+    target.upgrades = target.upgrades || {};
+    if (!target.upgrades[id]) {
+      const def = getUpgradeDefinition(id);
+      target.upgrades[id] = structuredClone(def?.defaultState || {});
+    }
+    return target.upgrades[id];
+  }
 }
 
+const stateManager = new StateManager();
+
+export default stateManager;
+
+export const createEmptyDailyMetrics = (...args) => stateManager.createEmptyDailyMetrics(...args);
+export const ensureDailyMetrics = (...args) => stateManager.ensureDailyMetrics(...args);
+export const ensureStateShape = (...args) => stateManager.ensureStateShape(...args);
+export const buildBaseState = (...args) => stateManager.buildBaseState(...args);
+export const buildDefaultState = (...args) => stateManager.buildDefaultState(...args);
+export const initializeState = (...args) => stateManager.initializeState(...args);
+export const replaceState = (...args) => stateManager.replaceState(...args);
+export const getState = (...args) => stateManager.getState(...args);
+export const getHustleState = (...args) => stateManager.getHustleState(...args);
+export const getAssetState = (...args) => stateManager.getAssetState(...args);
+export const countActiveAssetInstances = (...args) => stateManager.countActiveAssetInstances(...args);
+export const getUpgradeState = (...args) => stateManager.getUpgradeState(...args);
+
+export {
+  configureRegistry,
+  getRegistrySnapshot,
+  getHustleDefinition,
+  getAssetDefinition,
+  getUpgradeDefinition,
+  getMetricDefinition
+};
+export { createAssetInstance, normalizeAssetInstance, normalizeAssetState } from './state/assets.js';
+export { ensureNicheStateShape } from './state/niches.js';
